feat: separates the existing code into modules and jest setup
fix #10
